requirements.txt
Babel==2.5.3
beautifulsoup4==4.6.0
brotlipy==0.7.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
croniter==0.3.20
Django==2.0.4
django-anymail==2.0
django-cors-headers==2.2.0
django-environ==0.4.4
django-filter==1.1.0
django-htmlmin==0.10.0
django-ipware==2.0.1
django-money==0.13
django-phonenumber-field==2.0.0
django-rq==1.0.1
django-starterkit==1.0.3
django-tenants==2.0.0
django-trapdoor==1.0.8
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-msgpack==1.0.2
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
msgpack-python==0.5.6
phonenumberslite==8.9.3
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py-moneyed==0.7.0
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
redis==2.10.6
requests==2.18.4
rq==0.10.0
rq-scheduler==0.8.2
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
